fix: move coingecko entities to server.ts, misprint fix
