+++ app/entities/coingecko/coingecko-coins-schema.ts
