+++ app/entities/coingecko/coingecko-token-info-schema.ts
